Refactoring the diffing code and fixing type errors
